requirements.txt
Flask
google-cloud-speech
google-cloud-translate
google-cloud-texttospeech
grpcio
gunicorn